Attempt to draw something

+++ utopia/src/system/snes/ppu/cgram.rs
@@ -20,7 +20,7 @@ impl Cgram {
         }
     }
 
-    pub fn color(&mut self, index: u8) -> u16 {
+    pub fn color(&mut self, index: u16) -> u16 {
         self.data[index as usize]
     }
